several import fixes
